+++ front/src/app/pharmacy/type/type.component.ts
@@ -86,14 +86,16 @@ uploadFile(event){
 }
 
 onDelete(id: string) {
+  if(confirm('This can\'t be revert after deleted')){
 
-  this.Jarwis.deleteType(id).subscribe(  
+    this.Jarwis.deleteType(id).subscribe(  
+        
+      data => this.handleResponse(data),
+      error => this.handleError(error), 
       
-    data => this.handleResponse(data),
-    error => this.handleError(error), 
-    
-  );
+    );
   }
+}
 
 
   onSubmit(form: NgForm) {
@@ -108,7 +110,7 @@ onDelete(id: string) {
   }
 
   handleResponse(data) {    // 
-    let snackBarRef = this.snackBar.open("Added successfully", 'Dismiss', {
+    let snackBarRef = this.snackBar.open("Operation successfully", 'Dismiss', {
       duration: 2000
     })   
     this.router.navigateByUrl('/Admin/(side:item_type)');
